CQL: dissect Prepared response
It was not really dissected thus far.
Since its result metadata is identical to rows result, I've extracted it to its own function and used in both.

It also fixes a subtle bug in result parsing, while at it:
if Has_more_pages flag is set, then paging_state field should be the first to apper in the metadata, before the (optional) global_table_spec

Signed-off-by: Yaniv Kaul <mykaul@gmail.com>