Qt: Move the I/O Graph tracer to frame selected in main window
Clicking on the I/O Graph can select a frame in the main window.
Implement the reverse operation; when selecting a frame in the new
window, move the I/O Graph tracer to the appropriate location on
the current graph. (This is not an invertible operation because the
points on the graph are bucketed in time.)

As part of implementing this, separate out the part of the mouseMoved()
function that updates the graph hint from the parts that update state
necessary for the hint. Replace the calls to mouseMoved(NULL) used just to
update the hint to the new updateHint() function, for clarity. This
removes the possiblity of mouseMoved() being called with a nullptr
QMouseEvent.

Fix #12909