QUIC: Restore cipher after a key phase change if decryption fails
If a short header, after removing header protection, indicates a
key phase change, but payload decryption fails, discard the
new cipher and go back to the previous one. Only save the new
cipher if decryption succeeds.

This prevents problems where, e.g., padding is interpreted as a
possible short header packet (quite likely when a zero length
connection ID is used), indicates KP1 (as it will about half the time),
fails, and then the subsequent real packets are KP0. This is not the
third phase but the true original phase; we don't want to move on to
yet another new key.

We can (and might) put in extra tests for neqo-style padding outside
the QUIC packet with a zero length connection ID, but we have to do
this to have a hope of dealing with truly random padding, if it ever
exists.

Related to #16915