Qt: Fix flow graph scroll direction via keyboard
The y (time) axis is range reversed. QCP reversed ranges still
have the mathematically smaller element in lower and the larger
in upper, so reverse the direction of movement in panAxes.

Also fix the keyboard right arrow allowing the graph to be scrolled
to right farther than the QScrollBar allows, into extra empty space.

Fix #12932