Bluetooth: Improve packet dissection when PDU type is unknown
This commit allows us to dissect periodic advertising and
Broadcast ISO data even when the PDU type is unknown.
This is useful when analyzing pcap files generated from
raw packet data only containing access addresses, channel and
raw payload.

To distinguish between connections and periodic advertisers,
periodic advertisers needs to be stored in a new periodic_adv_info_tree.
With this it is possible to set the PDU type to
BTLE_PDU_TYPE_ADVERTISING for periodic advertisers.

For now these entries are kept to be of type connection_info_t,
even though this doesn't really make sense for periodic advertisers.
Some more cleanup is needed here.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>