Qt: Move I/O Graph reset button to button box
There have been a number of controls added to the I/O Graph
dialog recently, which increases its minimum width. Move the
Reset button to the QDialogButtonBox, using the standard button.
There's a lot of empty space in the button box due to all the
controls in the line above, and this more efficient use of space
allows the dialog to be shrunk a little bit if desired.