DTLS: Support heuristics for DTLS 1.3 unified header
Support heuristics for the DTLS 1.3 unified header, which uses
a a range of reserved values for the first oct that are not exactly
content type (chosen to work with the RTP/RTCP/STUN/TURN/ZRTP/DTLS
multiplexing scheme in RFC 7983). With the unified header, the
legacy_record_version bytes are encrypted, so the heuristic is
weaker. Only pass the heuristic if we've already detected a DTLS
session on the connection (with a matching CID if the CID bit is
present.)

Note that if we haven't detected the session, we won't be able to
decrypt it anyway.

Related to #18071, #19736.