LAPD: Set source address
The destination and source address are both being set to the
destination value. This causes the normal all source Flow Graph
(as opposed to the one created by the VoIP Calls Dialog) to have
empty lines.

Set the dl_src and dl_dst addresses using the AT_STRINGZ address
instead of col_set_str so that it works for both the Resolved and
Unresolved columns, and set the src and dst addresses as shallow
copies of those.

Fixup 7ac827fd74cf7b1eac80d54b5ce0001f15a587e3