sequence analysis: Fix inconsistent and illogical use of "Any"
The sequence analysis "any_addr" element says that it determines
whether any address (DL or network) is used versus network addresses
only.

However, the logic is clearly reversed; when any_addr is true, only
network addresses are used, and when it is false, the last set address
is used (and thus if only a data link address is set, no address is
used.) Confusingly, when the SequenceDialog is first opened from the
"Flow Graph" menu item, the any_addr member defaults to false, but
the index in the addresssComboBox defaults to saying "Any". However,
after initial setup, if the combobox is changed, then "Network"
causes "any_addr" to be false and all addresses displayed, whereas
"Any" causes "any_addr" to be true but only network addresses displayed.

Manually set up the combobox and use associated user data to ensure
that these stay consistent. Reverse the internal handling so that
any_addr means what you think it means - this also fixes the
tshark flow tap.

Fix #19818