Qt: Check Flow Graph "Limit to Display Filter" in constructor
The documentation for the Flow Graph claims that the checkbox
will be checked if the main display filter is active before the
window is opened. Make that actually the case, since most other
dialogs do that. (See #19054.)

Connect the displayFilterCheckBox toggled signal manually afterwards,
so that an extra tap isn't triggered by enabling it in the constructor.