wimax: move shared preferences to a header file.
Don't declare them privately in each file that uses it; that's a recipe
for disaster, as, for example, a change to the type of the preference in
the file that defines it, without corresponding changes to all files
that declare it, won't be detected.

Fixes #19822, which is the result of such a change.