kafka: Dissect assuming a supported API version
If the API version in the Kafka request is not for a supported
version, but we do support dissecting it for a different version,
dissect it according to the nearest supported version, adding an
expert info to inform about the assumption.

This prevents some infinite loops on fuzzed data where the dissector
was dissecting nothing and not advancing the offset because the
API version was too low for the message. Most (all?) of these are where
the API version, a signed integer according to the Kafka protocol, was
less than 0. A different option would be to just abort dissecting the
message entirely on a negative or unsupported API version, but as
bit errors are always possible doing something seems reasonable.

Fix #19768