Qt: Add context menu to launch I/O Graph
Add launching an I/O graph for a field to the ProtoTree context menu.
For a particular field type, add all the calculation types that it
supports. Only include the "advanced" calculations that are relevant
for individual fields.

We might want to add actions for the entire packet, for use from
PacketList as well.

All IOGraphDialogs for open files share the same UAT. Look for an
dialog with an open file, and add the graph if there is one. If not,
open a new dialog.

Fix #11362