Qt: Fix I/O Graphs crash with multiple dialogs
The UatModel accesses the uat_t's raw data directly, but also
keeps a list of dirty and changed records. That doesn't work
when more than one UatModel is accessing the same table and
one edits the uat. The other model ends up with inconsistent
information in those lists and crashes with illegal indices.
At least for now, use a single I/O Graph UatModel parented to
the main application, created when the iog uat is created.
Have the dialog views all share that model.

To keep the list of IOGraphs in sync with the UAT, rely on the
models signals about rows being inserted, removed, moved, or
the model reset.

Make sure to reload the UatModel if the uat is reloaded outside
the model; this prevents data inconsistency and crashing when
the profile is changed while the I/O Graphs window is opened.

Fix #15815. Related to #16373. Related to #13855 (using the
signals makes it easier to later implement drag and drop.)