wslua: Use Lua_Integer where we can for the wslua_ functions
We can assume that Lua has at least 32 bit integers. Use the conversion
functions with integers for 32 bit integers.

On Lua 5.3, this avoids some extra conversions to and from the floating
point Lua_Number type.

On Lua 5.2 and earlier, use the unsigned conversion functions when
doing an unsigned conversion, because they're not always equivalent
(e.g. on Windows x86, conversion from a Lua_Number uses inline assembly,
and if integers are 32 bit, only provides a 32 bit register to put the
result of FISTP into, which fails for large unsigned values. It only
provides a 64 bit register if you use the unsigned conversion functions.
(#18367).

Converting from a negative floating point to unsigned is undefined
behavior, which does what you might expect on x86 platforms (wraparound
conversion, same as if converting to the signed integer type and then
to unsigned) and does not on ARM platforms. So again it's important
to choose the correct conversion function.
(Commit 15392c324d5eaefcaa298cdee09cd5b40b12e09c)