Qt: Add visual notification to UAT Model entries that won't be saved
A new UAT entry appended via UatModel::appendEntry is not initially
marked as valid. It can only be marked as valid after some values
are altered by the user. Only valid entries are accessible from
dissectors or saved to file.

The I/O Graph Dialog, when it adds entries, including default graphs
for the currently applied filter, adds entries that do not have any
invalid fields (and can be immediately graphed) that get a pink
backround, but will not be saved until some field is altered and
the record changes to validated.

This generally makes sense, but to aid the user, show the temporary
records that will not get saved with an italicized font (or non
italicized if, for some reason, the user's default font is italicized.)
This provides a visual cue of a temporary record.