Add a ENC_TIME_ZBEE_ZCL time encoding and use it.
Add an epoch value for the Zigbee ZCL epoch - 2000-01-01 00:00:00 UTC -
to wsutil/epochs.h, and use that.

Add a time stamp format consisting of a 4-byte or 8-byte count of
seconds since that epoch, ENC_TIME_ZBEE_ZCL, and use that instead of
manually extracting the value, constructing an nstime_t, and adding that
value.