lua: Deregister heuristic lists added via heuristic_new on reload
Heuristic dissectors lists don't necessarily (and don't usually)
have the name of a given dissector handle.

There's still a leak when reloading plugins because the created
heuristic dissector list isn't returned to the caller. That also
means there's no way of knowing if it succeeded, and using the
heuristic list involves calling proto:register_heuristic with
the name, but rather than calling a function on a returned
object.

Fix #19603