smb2: fix ahndling of pre-3.10 Negotiate Protocol.
Prior to 3.10:

Negotiate Protocol requests didn't include negotiate contexts, and the
NegotiateContextOffset, NegotiateContextCount, and Reserved2 fields
were, instead, a must-be-zero and must-be-ignored ClientStartTime field.

Negotiate Protocol responses didn't include negotiate contexts, and the
NegotiateContextCount and NegotiateContextOffset fields were, instead, a
Reserved and Reserved2 field, respectively, both of which must be zero
and must be ignored.

If 3.10 or later aren't offered in a Negotiate Protocol request, and the
NegotiateContextOffset or NegotiateContextCount fields are non-zero,
report that as an error, and force the offset and count values to 0, so
that we don't try to dissect the contexts.

If 3.10 or later isn't selected in a Negotiate Protocol reqply, display
Reserved and Reserved2 fields rather than NegotiateContextCount and
NegotiateContextOffset fields and, if they have non-zero values, report
that as an error, and force the value to 0, so that we don't try to
dissect the contexts.

This fixes some dissection errors.  (We also should check whether the
offset value goes beyond the end of the packet. and flag that as well.)

Update some docs.microsoft.com URLs to learn.microsoft.com while we're
at it.