Qt: Fix RTP Player crash
marker_stream_ might not exist when playFinished is called.
In particular, deleting the RtpAudioStreams might cause
playFinished to be called when the audio streams delete
their audio output in certain cases, such as if they are
paused. (The behavior might be slightly different on Qt6,
where QAudioSink is used vs. Qt5 where QAudioOutput is used.)

Fix #19596