CMake: Have CheckAPI targets depend on the source files
This ensures that generated source files are present and up to date.
It allows running the checkAPI target on a clean source tree, instead
of after a build.

Get rid of writing the dissector file list to a file and pass it
in to the CMake command as a parameter so that it's easier to
depend on the files.