MySQL: Handle displaying parameters and result sets in charset
Translate collation / charset types to Wireshark encodings that
are supported.

Store the character_set_client and character_set_results for a
connection. Handle them when the system variables are changed
in response to a SET NAMES or similar. Store them as per-packet
state because they can change during the connection.

When we have the charset number from a field packet, use that
for the field. Otherwise fallback to the character_set_client
or character_set_results as appropriate. We always use
character_set_client for parameters and queries, and
character_set_results for errors.

Split result set strings from a FT_UINT_STRING to a FT_UINT24
and a FT_STRING, because the string length is always LITTLE_ENDIAN,
but the encoding could theoretically (for results, if not
chararacter_set_client) be Big Endian UTF-16 or UTF-32, and
FT_UINT_STRING doesn't support one endian for the length and another
for the text.

Default to UTF-8. Perhaps we could add a preference for a default.

Fix #19839