Const-ify dcerpc_sub_dissector structures.
This moves about 56 kBytes of data from a read-write data section to read-only.

The various dissectors were _not_ regenerated for this commit.  Instead, this
commit was produced using the following command:

   find * -type f -exec sed -i -e 's/\(["'\'']\|^\)static dcerpc_sub_dissector/\1static const dcerpc_sub_dissector/' \{\} \;

This updates the generator script and its test, as well as making an identical
change to all of the dissectors, regardless of whether or not those dissectors
were produced by the generator script.