TCP Stream: Make sure the SACK rel/abs sequence number is correct
Commit f255f6c683a1f3b895f3adf692505a79baddae34 switched the
tap information for the SACK blocks from always being relative
sequence numbers to being either absolute or relative depending
on the TCP dissector preference.

For now, to make the graph correct and allow switching between
relative and absolute sequence numbers, store the raw sequence
number information for each segment.

This is cheap enough to handle (we only need to examine two
segments to ensure that we get both directions of base seq, since
the SYN doesn't give us the reverse direction) that we could do it
in the tap and store it at the graph level to save memory.

Fixup 24a0417eb05e0edb979c3ff004003167b46c269d