uat: Minor updates
Add PT_TXTMOD_DISSECTOR to the casees handled by uat_fld_tostr.
Reorder the cases to more closely match putfld, which uat_save
uses.

Get rid of an unnecessary strdup of a GString.

uat_load_str dupes the string entry passed in, so that can be const.

Note that uat_fld_tostr doesn't do what the documentation says,
though sharkd_session uses it as-is and puts the results in JSON
instead of the UAT CSV format.