docs: Update Display Filter Macros, UATs in WSUG and elsewhere
Add Display Filter Expressions to the Analyze menu items table

Update the sections on Display Filter Macros to reflect that it is
no longer a UAT and the new filename.

Move the Display Filter Macros syntax to the section on defining
and saving Display Filter Macros to consolidate information on macros.

Indicate which UATs are per profile and which apply to all profiles.
Add cross references.

Put the UAT sections under the main UAT section.

Part of #16982