Qt: activateWindow when asking about unsaved packets
Wireshark might not be the active application when testing if the
capture file can be closed. E.g., on Windows, multiple Wireshark
processes can be open, another application have focus, and the user
right-click on the Wireshark icon and select "Close all windows."

When there's only a single Wireshark process running this activates
Wireshark and shows the modal dialog on top as expected, but for
multiple processes. Explicitly call activateWindow on the main window
before calling exec() on the QMessageBox.

Note that this still doesn't bring the window to the top of the desktop,
but instead flashes the taskbar. That's intentional Windows design:
https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-setforegroundwindow

In Qt 5 it was possible to override the platform default
https://doc.qt.io/qt-5/qwindowswindowfunctions.html#WindowActivationBehavior-enum
but that doesn't seem to be ported to Qt 6 yet
https://www.qt.io/blog/platform-apis-in-qt-6

Fix #16309