capture: Always pass the interface description to the capture child
Interfaces are passed from Wireshark or tshark to dumpcap using the
name, which avoids having to retrieve the interface list. We already
pass the description for extcaps; pass it for other interfaces as well
to put in the IDB and use for the filename.

This make captures spawned from Wireshark or tshark have the same
interface description and same tempfile name as when running dumpcap
directly. It particularly affects captures using the loopback interface,
on both Windows (where a Win32 API call cannot retrieve the Loopback
interface description) and other OSes.

Fix #18466