epan: Rename tvb_[child_]uncompress to tvb_[child_]uncompress_zlib
It's confusing to have a function just called "uncompress" that
specifically only does zlib/deflate decompression. Have the current
names still work by calling the old code, but deprecate them so there
can be a warning when compiling.

Deprecating the functions allows the possibility of having a
tvb_uncompress function that takes the algorithm as a parameter in
some future version.

Do the same with TvbRange:uncompress[_zlib] in Lua.

Part of #16451