Qt: Raise autostop filesize limit to 2 TB, enforce in the GUI
The variable used to control the autostop filesize (or filesize
at which to switch to the next file in multiple file mode) is a
32 bit unsigned integer, but it measures kB. Therefore we can easily
allow larger sizes. There is a small limitation because the QSpinBox
uses a signed integer (we could evade that by allowing larger sizes
only with MB or GB selected), but even with that, it's easy to support
2 TB, which is quite a bit larger.

We have large file support everywhere (tshark, Wireshark, libpcap,
etc.) so the size alone shouldn't be a problem. Such a very big
file might run into issues from being more than 2^31 frames in the
GUI (QTreeView uses an int for the number of rows) or 2^32 frames
viewed in tshark, but that would equally happen with *no* filesize
limit.

The autostop options, like the ring buffer options, must be greater
than zero. The GUI was only enforcing the latter. Set the minimums
for the spinboxes.

Also set maximums for the file size spinboxes, rather than popping up
a warning message when the user puts in a disallowed value. Change
the maximums depending on whether kB, MB, or GB are selected.

Fix the warning message (which shouldn't happen anymore), which was
incorrectly using binary prefixes even though the enforced maximum
value is actually a decimal/metric prefix.

Fix #5691